// src/common/enums/NutritionUnit.ts
export enum ENutritionUnit {
  CALORIES = 'cal',
  PROTEIN = 'g',
  CARBOHYDRATES = 'g',
  FAT = 'g',
  SATURATED_FAT = 'g',
  POLYUNSATURATED_FAT = 'g',
  MONOUNSATURATED_FAT = 'g',
  TRANS_FAT = 'g',
  FIBER = 'g',
  SUGAR = 'g',
  SODIUM = 'mg',
  POTASSIUM = 'mg',
  CHOLESTEROL = 'mg',
  VITAMIN_A = 'IU',
  VITAMIN_C = 'mg',
  CALCIUM = 'mg',
  IRON = 'mg',
}